Renamed "interactions" -> "operations"
