+++ pink/rondis/string/db_operations.h
